+++ src/main/kotlin/ru/netology/interfaces.kt
@@ -1,3 +1,4 @@
+package ru.netology
 /*interface Attachment {
     val type: String
 }
